sensors.yaml added
Sensors.yaml adds a custom sensor to calculate the potential maximum solar yield based on the weather station's measured radiation in W/m2, and calculated to adjust for the total panel area, and efficiency factor of the panels. Comments in sensors.yaml file. The mini graph on the Solar dashboard tab was also adjusted.