Our Groceries Shopping list and Netflix RSS feed added
Got the Our Groceries card to finally display (it was not showing under HACS frontends when I filtered to find it, but a complete reading of the list found it) although the actual list is not showing yet. Also added an RSS feed to show what titles are leaving Netflix soon.