Time left on battery fixes
Fixed sensor template to ignore zero current (would have given errors with division by zero) and charging current.

Instead for those conditions it now uses a calculation based on current SOC for a 8 A discharge current (ie. light load). So as the battery charges this will show an increase in time, and as soon as it starts to discharge, it will show a correct time to go. Another fix was to ensure that the projected time to go down to Minimum SOC, stops with a  reading of 0 hours when it reaches that level, or below Min SOC.

Details on the comments in sensors.yaml file.