Add files via upload
Found that at a very low current draw (just below 0 Amps) the calculation was showing unrealistic time to go of over 1,000 hours). Strictly speaking this is quite correct, but in fact the actual current load is about 500 Watts. This was happening when the battery was fully charged, so won't be an issue if it is actually discharging properly. But to make it more real, I continued to apply the projected load from AC Loads for charging, as well as now to anything between 0 Amps to -1 A.