Time to Go on AC Loads
Still had some issues with battery current less than -1 Amp draw throwing the numbers out as it changed to just below -1. Realised that if Energy in Watts on AC Load and Energy in Watts on battery are about the same, I may as well just use the AC Load to calculate the estimated load if the battery switched exclusively to battery. It will provide a more consistent result. Original code is still commented in sensors.yaml file.