Person tracking changed
Instead of tracking a specific device for location, I changed that to rather track a person as an entity. This means that a person can have their devices switched around or changed as the device which defines their location. This make it more generic and flexible than tracking the device itself for automations.