Update sensors.yaml
Added a custom sensor based on the AC Loads modbus sensor, because you cannot set/change an icon for modbus sensors. This custom sensor displays the same value but the icon will change depending on that value. This can be done in custom cards on the dashboard, but I'm pulling in HA data into my Stream Deck device and now the icon can also change based on the value.